adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "AI4ES ADK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk (>=1.12.0,<2.0.0)",
    "litellm (>=1.72.2,<2.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-core (>=0.3.66,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "qdrant-client (>=1.14.3,<2.0.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "fastembed (>=0.7.1,<0.8.0)",
    "google-genai",
    "psycopg2-binary",
    "asyncpg>=0.29.0,<1.0.0",
    "boto3>=1.40.15",
    "gunicorn>=20.1.0,<22.0.0",
    "langfuse>=3.3.5",
    "openinference-instrumentation-google-adk>=0.1.4",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
